fix(package): Fix module package and install
