+++ setup.cfg
@@ -24,4 +24,4 @@ where=src
 
 [options.entry_points]
 console_scripts =
-    pvecontrol = pvecontrol.pvecontrol:main
+    pvecontrol = pvecontrol:main
