+++ src/pvecontrol.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python3
+
+import re
+import sys
+from pvecontrol import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
