+++ src/pvecontrol/__main__.py
@@ -1,0 +1,4 @@
+import pvecontrol
+
+if __name__ == "__main__":
+    pvecontrol.main()
